server/requirements.txt
astroid==2.4.2
binwalk==2.2.0
blinker==1.4
breezy==3.0.2
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==5.0.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.7
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.0.0
dulwich==0.19.15
entrypoints==0.3
fastimport==0.9.8
Flask==1.1.1
gpg===1.13.1-unknown
hidpidaemon==18.4.6
httplib2==0.14.0
idna==2.8
isort==5.6.2
itsdangerous==1.1.0
Jinja2==2.11.2
kernelstub==3.1.4
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
Pillow==7.0.0
pop-transition==1.0.1
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pydbus==0.6.0
pyflatpak==0.4.0
PyGithub==1.43.7
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==2.4.6
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
pyRFC3339==1.1
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
repoman==1.2.2
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.3.3
screenkey==0.9
SecretStorage==2.3.1
sessioninstaller==0.0.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
supervisor==4.1.0
systemd-python==234
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
vboxapi==1.0
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.11.2
xkit==0.0.0
